+++ lib/active_encryption.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'active_encryption/version'
+require 'active_encryption/encryption_setting/yaml_store'
 
 module ActiveEncryption
   # The ActiveEncryption::Error is the base class for ActiveEncryption errors.