+++ CHANGELOG.md
@@ -6,4 +6,5 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+- Add encryption setting records and YAML-based store to persist them
 - Basic gem skeleton